feat(timesheet): enhance timesheet entry page with modular architecture
This commit refactors the timesheet entry page by modularizing
the JavaScript codebase. Key changes include:

- Splitting large scripts into modular components (e.g., grid, state, modals).
- Implementing modular handling for paid absences and job filtering.
- Adding a centralized summary update mechanism.
- Improving maintainability with reusable utilities and focused logic.

These changes improve code readability, maintainability, and
scalability while preserving all core functionalities. The
new modular structure lays the groundwork for easier feature
enhancements and bug fixes in the future.
