+++ workflow/static/js/timesheet/timesheet_entry/timesheet_autosave.js
@@ -1,3 +1,5 @@
+import { rowStateTracker } from "./state.js";
+
 let deletedEntries = [];
 
 /**
