+++ workflow/templates/time_entries/timesheet_entry.html
@@ -96,7 +96,7 @@
     {% block extra_js %}
     <!-- AG Grid Script -->
     <script src="{% static 'ag-grid-community/ag-grid-community.min.js' %}"></script>
-    <script src="{% static 'js/deserialize_timesheet.js' %}"></script>
-    <script src="{% static 'js/timesheet_autosave.js' %}"></script>
-    <script src="{% static 'js/timesheet_entry.js' %}"></script>
+    <script src="{% static 'js/timesheet/deserialize_timesheet.js' %}"></script>
+    <script src="{% static 'js/timesheet/timesheet_entry/timesheet_autosave.js' %}"></script>
+    <script type="module" src="{% static 'js/timesheet/timesheet_entry/timesheet_entry.js' %}"></script>
     {% endblock %}
